Split out multi_buffer transactions into separate module
